Split up pytransform3d.plot_utils
