+++ .coveragerc
@@ -9,7 +9,7 @@ omit =
 
 [report]
 omit =
-    pytransform3d/plot_utils.py
+    pytransform3d/plot_utils/*.py
     pytransform3d/editor.py
     pytransform3d/visualizer.py
 exclude_lines =
